Changing Joystick widget event to use a flot in the -1..1 range.

+++ DroidPlanner/src/com/droidplanner/widgets/joystick/JoystickMovedListener.java
@@ -4,7 +4,7 @@
 package com.droidplanner.widgets.joystick;
 
 public interface JoystickMovedListener {
-	public void OnMoved(int pan, int tilt);
+	public void OnMoved(double pan, double tilt);
 	public void OnReleased();
 	public void OnReturnedToCenter();
 }
